+++ src-tauri/src/commands/playback.rs
@@ -30,12 +30,6 @@ pub fn preload_stream<R: tauri::Runtime>(app_handle: tauri::AppHandle<R>, stream
 }
 
 #[tauri::command]
-pub fn set_bit_perfect_enabled<R: tauri::Runtime>(app_handle: tauri::AppHandle<R>, enabled: bool) -> Result<(), String> {
-    get_player(&app_handle)?.set_bit_perfect_enabled(enabled);
-    Ok(())
-}
-
-#[tauri::command]
 pub fn set_visualizer_enabled<R: tauri::Runtime>(app_handle: tauri::AppHandle<R>, enabled: bool) -> Result<(), String> {
     get_player(&app_handle)?.set_visualizer_enabled(enabled);
     Ok(())